Added some cam stuff
I added some of the code for the camera from lab 3, for some reason my app won't run because it needs to sync with gradle and I can't find a way to do that. So if you guys can try to run it and see if t works (or crashes) that would be great.