Added task deletion & linked subtasks->tasks
For deleting tasks I made it so that the tasks are created with checkboxes beside them that the user can check off and click the big red delete button to delete.

To link the subtasks and the tasks I had to create a separate class called Taskadapter that handled a lot of the functions for the tasks like keeping track of how many there are, removing them and navigating to the subtasks. Users can simply click the task and it will take them to the subtasks.